Add MpesaCallbackMixin for shared callback processing
