+++ mpesakit/mpesa_client.py
@@ -1,6 +1,7 @@
 """MpesaClient: A unified client for M-PESA services."""
 
 from mpesakit.auth import TokenManager, AsyncTokenManager
+from mpesakit.callback_mixin import MpesaCallbackMixin
 from mpesakit.http_client import MpesaHttpClient, MpesaAsyncHttpClient
 from mpesakit.services import (
     B2BService,
@@ -28,7 +29,7 @@ from mpesakit.services import (
 )
 
 
-class MpesaClient:
+class MpesaClient(MpesaCallbackMixin):
     """Unified client for all M-PESA services."""
 
     def __init__(
@@ -103,116 +104,8 @@ class MpesaClient:
             http_client=self.http_client, token_manager=self.token_manager
         )
 
-    def process_stk_callback(self, payload):
-        """Process STK Push callback payload.
 
-        Validates and parses STK Push simulation callback data.
-        """
-        from mpesakit.mpesa_express.schemas import StkPushSimulateCallback
-
-        return StkPushSimulateCallback.model_validate(payload)
-
-    def process_stk_query_callback(self, payload):
-        """Process STK Push query response payload.
-
-        Validates and parses STK Push query response data.
-        """
-        from mpesakit.mpesa_express.schemas import StkPushQueryResponse
-
-        return StkPushQueryResponse.model_validate(payload)
-
-    def process_account_balance_callback(self, payload):
-        """Process account balance callback payload.
-
-        Validates and parses account balance query result callback.
-        """
-        from mpesakit.account_balance.schemas import AccountBalanceResultCallback
-
-        return AccountBalanceResultCallback.model_validate(payload)
-
-    def process_account_balance_timeout(self, payload):
-        """Process account balance timeout callback payload.
-
-        Validates and parses account balance query timeout notification.
-        """
-        from mpesakit.account_balance.schemas import AccountBalanceTimeoutCallback
-
-        return AccountBalanceTimeoutCallback.model_validate(payload)
-
-    def process_b2c_callback(self, payload):
-        """Process B2C (Business-to-Customer) callback payload.
-
-        Validates and parses B2C payment result callback.
-        """
-        from mpesakit.b2c.schemas import B2CResultCallback
-
-        return B2CResultCallback.model_validate(payload)
-
-    def process_b2b_callback(self, payload):
-        """Process B2B Express Checkout callback payload.
-
-        Validates and parses B2B Express Checkout response data.
-        """
-        from mpesakit.b2b_express_checkout.schemas import B2BExpressCheckoutCallback
-
-        return B2BExpressCheckoutCallback.model_validate(payload)
-
-    def process_transactions_callback(self, payload):
-        """Process transaction status callback payload.
-
-        Validates and parses transaction status query result callback.
-        """
-        from mpesakit.transaction_status.schemas import TransactionStatusResultCallback
-
-        return TransactionStatusResultCallback.model_validate(payload)
-
-    def process_bill_manager_callback(self, payload):
-        """Process bill manager callback payload.
-
-        Validates and parses bill manager payment notification.
-        """
-        from mpesakit.bill_manager.schemas import BillManagerPaymentNotificationRequest
-
-        return BillManagerPaymentNotificationRequest.model_validate(payload)
-
-    def process_tax_remittance_callback(self, payload):
-        """Process tax remittance callback payload.
-
-        Validates and parses tax remittance result callback.
-        """
-        from mpesakit.tax_remittance.schemas import TaxRemittanceResultCallback
-
-        return TaxRemittanceResultCallback.model_validate(payload)
-
-    def process_dynamic_qr_code_callback(self, payload):
-        """Process dynamic QR code callback payload.
-
-        Validates and parses dynamic QR code generation response.
-        """
-        from mpesakit.dynamic_qr_code.schemas import DynamicQRGenerateResponse
-
-        return DynamicQRGenerateResponse.model_validate(payload)
-
-    def process_ratiba_service_callback(self, payload):
-        """Process Ratiba (Standing Order) callback payload.
-
-        Validates and parses M-PESA Ratiba services callback.
-        """
-        from mpesakit.mpesa_ratiba.schemas import StandingOrderCallback
-
-        return StandingOrderCallback.model_validate(payload)
-
-    def process_reversal_callback(self, payload):
-        """Process transaction reversal callback payload.
-
-        Validates and parses transaction reversal result callback.
-        """
-        from mpesakit.reversal.schemas import ReversalResultCallback
-
-        return ReversalResultCallback.model_validate(payload)
-
-
-class AsyncMpesaClient:
+class AsyncMpesaClient(MpesaCallbackMixin):
     """Unified async client for all M-PESA services."""
 
     def __init__(
@@ -285,111 +178,3 @@ class AsyncMpesaClient:
         self.ratiba = AsyncRatibaService(
             http_client=self.http_client, token_manager=self.token_manager
         )
-
-    def process_stk_callback(self, payload):
-        """Process STK Push callback payload.
-
-        Validates and parses STK Push simulation callback data.
-        """
-        from mpesakit.mpesa_express.schemas import StkPushSimulateCallback
-
-        return StkPushSimulateCallback.model_validate(payload)
-
-    def process_stk_query_callback(self, payload):
-        """Process STK Push query response payload.
-
-        Validates and parses STK Push query response data.
-        """
-        from mpesakit.mpesa_express.schemas import StkPushQueryResponse
-
-        return StkPushQueryResponse.model_validate(payload)
-
-    def process_account_balance_callback(self, payload):
-        """Process account balance callback payload.
-
-        Validates and parses account balance query result callback.
-        """
-        from mpesakit.account_balance.schemas import AccountBalanceResultCallback
-
-        return AccountBalanceResultCallback.model_validate(payload)
-
-    def process_account_balance_timeout(self, payload):
-        """Process account balance timeout callback payload.
-
-        Validates and parses account balance query timeout notification.
-        """
-        from mpesakit.account_balance.schemas import AccountBalanceTimeoutCallback
-
-        return AccountBalanceTimeoutCallback.model_validate(payload)
-
-    def process_b2c_callback(self, payload):
-        """Process B2C (Business-to-Customer) callback payload.
-
-        Validates and parses B2C payment result callback.
-        """
-        from mpesakit.b2c.schemas import B2CResultCallback
-
-        return B2CResultCallback.model_validate(payload)
-
-    def process_b2b_callback(self, payload):
-        """Process B2B Express Checkout callback payload.
-
-        Validates and parses B2B Express Checkout response data.
-        """
-        from mpesakit.b2b_express_checkout.schemas import B2BExpressCheckoutCallback
-
-        return B2BExpressCheckoutCallback.model_validate(payload)
-
-    def process_transactions_callback(self, payload):
-        """Process transaction status callback payload.
-
-        Validates and parses transaction status query result callback.
-        """
-        from mpesakit.transaction_status.schemas import TransactionStatusResultCallback
-
-        return TransactionStatusResultCallback.model_validate(payload)
-
-    def process_bill_manager_callback(self, payload):
-        """Process bill manager callback payload.
-
-        Validates and parses bill manager payment notification.
-        """
-        from mpesakit.bill_manager.schemas import BillManagerPaymentNotificationRequest
-
-        return BillManagerPaymentNotificationRequest.model_validate(payload)
-
-    def process_tax_remittance_callback(self, payload):
-        """Process tax remittance callback payload.
-
-        Validates and parses tax remittance result callback.
-        """
-        from mpesakit.tax_remittance.schemas import TaxRemittanceResultCallback
-
-        return TaxRemittanceResultCallback.model_validate(payload)
-
-    def process_dynamic_qr_code_callback(self, payload):
-        """Process dynamic QR code callback payload.
-
-        Validates and parses dynamic QR code generation response.
-        """
-        from mpesakit.dynamic_qr_code.schemas import DynamicQRGenerateResponse
-
-        return DynamicQRGenerateResponse.model_validate(payload)
-
-    def process_ratiba_service_callback(self, payload):
-        """Process Ratiba (Standing Order) callback payload.
-
-        Validates and parses M-PESA Ratiba services callback.
-        """
-        from mpesakit.mpesa_ratiba.schemas import StandingOrderCallback
-
-        return StandingOrderCallback.model_validate(payload)
-
-    def process_reversal_callback(self, payload):
-        """Process transaction reversal callback payload.
-
-        Validates and parses transaction reversal result callback.
-        """
-        from mpesakit.reversal.schemas import ReversalResultCallback
-
-        return ReversalResultCallback.model_validate(payload)
