+++ stubs/mpesakit/callback_mixin.pyi
@@ -1,0 +1,13 @@
+class MpesaCallbackMixin:
+    def process_stk_callback(self, payload): ...
+    def process_stk_query_callback(self, payload): ...
+    def process_account_balance_callback(self, payload): ...
+    def process_account_balance_timeout(self, payload): ...
+    def process_b2c_callback(self, payload): ...
+    def process_b2b_callback(self, payload): ...
+    def process_transactions_callback(self, payload): ...
+    def process_bill_manager_callback(self, payload): ...
+    def process_tax_remittance_callback(self, payload): ...
+    def process_dynamic_qr_code_callback(self, payload): ...
+    def process_ratiba_service_callback(self, payload): ...
+    def process_reversal_callback(self, payload): ...
